requirements.txt
tweepy==3.5.0
PyYAML==3.12
Pillow==4.1.0
wget==3.2
